+++ src/glib/slist.rs
@@ -94,8 +94,8 @@ impl<T> SList<T> {
         }
     }
 
-    pub fn len(&self) -> uint {
-        unsafe { ffi::g_slist_length(self.pointer) as uint }
+    pub fn len(&self) -> usize {
+        unsafe { ffi::g_slist_length(self.pointer) as usize }
     }
 
     pub fn clear(&mut self) {
@@ -111,10 +111,10 @@ impl<T> SList<T> {
     }
 }
 
-impl<T> Index<uint> for SList<T> {
+impl<T> Index<usize> for SList<T> {
     type Output = T;
 
-    fn index<'a>(&'a self, _rhs: &uint) -> &'a T {
+    fn index<'a>(&'a self, _rhs: &usize) -> &'a T {
         self.nth(*_rhs as u32)
     }
 }
